refactor(foundation): move DriftResult, FitnessTestResult, consumer types (A13)
Move the contract-drift data types that consumers actually touch into
mockforge-foundation so they can be used without depending on deprecated
core modules:

- DriftResult (pure data; the from_diff_result helper stays in core as
  a free function drift_result_from_diff because it needs
  BreakingChangeDetector + ContractDiffResult)
- FitnessTestResult (pure data)
- AppType, ConsumingApp, SDKMethod, ConsumerMapping, ConsumerImpact
  (pure data; ConsumerMappingRegistry + ConsumerImpactAnalyzer stay in
  core)

Core's `contract_drift::{types,fitness,consumer_mapping}` re-export from
foundation so all existing call sites (engines, analyzers, registries)
keep working.

Enables mockforge-recorder/sync_drift.rs to import DriftResult from
foundation; the file keeps module-scoped allow(deprecated) for the
DriftBudgetEngine use (which depends on OpenApiSpec and stays in core).

All 42 foundation + 1818 core + 231 recorder tests pass; full workspace
clippy clean.
