+++ crates/mockforge-foundation/src/contract_drift_types.rs
@@ -256,3 +256,159 @@ impl BreakingChangeRule {
         }
     }
 }
+
+// ============================================================================
+// Fitness test result (pure data; evaluator logic stays in core)
+// ============================================================================
+
+/// Result of evaluating a fitness function
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct FitnessTestResult {
+    /// ID of the fitness function that was evaluated
+    pub function_id: String,
+    /// Name of the fitness function
+    pub function_name: String,
+    /// Whether the test passed
+    pub passed: bool,
+    /// Human-readable message about the result
+    pub message: String,
+    /// Metrics collected during evaluation
+    pub metrics: HashMap<String, f64>,
+}
+
+// ============================================================================
+// Consumer mapping types (pure data; analyzer logic stays in core)
+// ============================================================================
+
+/// Type of consuming application
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
+#[serde(rename_all = "snake_case")]
+pub enum AppType {
+    /// Web application
+    Web,
+    /// Mobile application (iOS)
+    #[serde(rename = "mobile_ios")]
+    MobileIos,
+    /// Mobile application (Android)
+    #[serde(rename = "mobile_android")]
+    MobileAndroid,
+    /// Internal tool or service
+    InternalTool,
+    /// CLI tool
+    Cli,
+    /// Other/unknown
+    Other,
+}
+
+impl std::fmt::Display for AppType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            AppType::Web => write!(f, "Web App"),
+            AppType::MobileIos => write!(f, "Mobile App (iOS)"),
+            AppType::MobileAndroid => write!(f, "Mobile App (Android)"),
+            AppType::InternalTool => write!(f, "Internal Tool"),
+            AppType::Cli => write!(f, "CLI Tool"),
+            AppType::Other => write!(f, "Other"),
+        }
+    }
+}
+
+/// A consuming application that uses SDK methods
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
+pub struct ConsumingApp {
+    /// Unique identifier for the app
+    pub app_id: String,
+    /// Human-readable name
+    pub app_name: String,
+    /// Type of application
+    pub app_type: AppType,
+    /// Optional repository URL
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub repository_url: Option<String>,
+    /// Timestamp when this app was last updated
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub last_updated: Option<i64>,
+    /// Optional description
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+}
+
+/// An SDK method that calls an endpoint
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
+pub struct SDKMethod {
+    /// SDK name (e.g., "typescript-sdk", "python-sdk")
+    pub sdk_name: String,
+    /// Method name (e.g., "getUser", "createOrder")
+    pub method_name: String,
+    /// List of consuming apps that use this SDK method
+    #[serde(default)]
+    pub consuming_apps: Vec<ConsumingApp>,
+}
+
+/// Mapping from endpoint to SDK methods
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ConsumerMapping {
+    /// Endpoint path or operation ID
+    pub endpoint: String,
+    /// HTTP method or protocol identifier
+    pub method: String,
+    /// SDK methods that call this endpoint/operation
+    #[serde(default)]
+    pub sdk_methods: Vec<SDKMethod>,
+    /// Timestamp when this mapping was created
+    #[serde(default)]
+    pub created_at: i64,
+    /// Timestamp when this mapping was last updated
+    #[serde(default)]
+    pub updated_at: i64,
+}
+
+/// Impact analysis result showing which consumers are affected by drift
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ConsumerImpact {
+    /// Endpoint path
+    pub endpoint: String,
+    /// HTTP method
+    pub method: String,
+    /// SDK methods that are affected
+    pub affected_sdk_methods: Vec<SDKMethod>,
+    /// Applications that are affected
+    pub affected_apps: Vec<ConsumingApp>,
+    /// Human-readable impact summary
+    pub impact_summary: String,
+}
+
+// ============================================================================
+// DriftResult (pure data; constructors / evaluator helpers stay in core)
+// ============================================================================
+
+/// Result of drift budget evaluation
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DriftResult {
+    /// Whether budget is exceeded
+    pub budget_exceeded: bool,
+    /// Number of breaking changes (definitely breaking)
+    pub breaking_changes: u32,
+    /// Number of potentially breaking changes (requires review)
+    #[serde(default)]
+    pub potentially_breaking_changes: u32,
+    /// Number of non-breaking changes
+    pub non_breaking_changes: u32,
+    /// Mismatches that are considered breaking (definitely breaking)
+    pub breaking_mismatches: Vec<Mismatch>,
+    /// Mismatches that are potentially breaking (requires review)
+    #[serde(default)]
+    pub potentially_breaking_mismatches: Vec<Mismatch>,
+    /// Mismatches that are non-breaking
+    pub non_breaking_mismatches: Vec<Mismatch>,
+    /// Current drift metrics
+    pub metrics: DriftMetrics,
+    /// Whether an incident should be created
+    pub should_create_incident: bool,
+    /// Results from fitness function tests
+    #[serde(default)]
+    pub fitness_test_results: Vec<FitnessTestResult>,
+    /// Consumer impact analysis (optional)
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub consumer_impact: Option<ConsumerImpact>,
+}
